requirements.txt
kopf==1.35.3
pykube-ng==21.10.0
cryptography==36.0.0
